Create multiple components, add Stripe button
