+++ client/my-sites/earn/customers/customer/style.scss
@@ -87,4 +87,8 @@
 			letter-spacing: -0.15px;
 		}
 	}
+
+	.customer__stripe-button {
+		margin-top: 30px;
+	}
 }
